Add ::Float64 to dt

+++ examples/bench_4p.jl
@@ -14,7 +14,7 @@ using ControlPlots
 set::Settings = deepcopy(load_settings("system.yaml"))
 
 # the following values can be changed to match your interest
-dt = 0.05
+dt::Float64 = 0.05
 set.solver = "DFBDF"              # IDA or DFBDF
 set.linear_solver = "GMRES"       # GMRES, LapackDense or Dense
 STEPS = 200
