+++ examples/initial_reel_out_4p.jl
@@ -7,7 +7,7 @@ using KiteUtils: Settings, load_settings
 set::Settings = deepcopy(load_settings("system.yaml"))
 
 # the following values can be changed to match your interest
-dt = 0.05
+dt::Float64 = 0.05
 set.solver = "DFBDF"   # IDA or DFBDF
 set.v_reel_out = 1.0   # initial reel-out speed [m/s]
 STEPS = 600
